AniLink typings completed
